タイマーページをPomodoroTimerコンポーネントとTimerControllerコンポーネントに分割
