+++ src/app/page.tsx
@@ -1,22 +1,26 @@
 "use client";
 
-import Button from "@/components/elements/Button";
-import { useState, useEffect } from "react";
-import { BsSkipEndFill } from "react-icons/bs";
-import { HiPause } from "react-icons/hi2";
-import { RiResetLeftFill } from "react-icons/ri";
-import { VscDebugStart } from "react-icons/vsc";
+import { useState } from "react";
+import PomodoroTimer from "./(timer)/_components/PomodoroTimer";
+import TimerController from "./(timer)/_components/TimerController";
 
+/** ポモドーロタイマー各数値の型定義 */
+interface TimerConfig {
+  focus: number;
+  shortBreak: number;
+  longBreak: number;
+  totalCycles: number;
+}
 /** ポモドーロタイマー初期値 */
-const defaultTimerConfig = {
+export const defaultTimerConfig: TimerConfig = {
   focus: 25 * 60,
   shortBreak: 5 * 60,
   longBreak: 30 * 60,
   totalCycles: 4,
-};
+} as const;
 
 /** ポモドーロタイマーフェーズ */
-type Phase = "focus" | "short-break" | "long-break";
+export type PomodoroTimerPhase = "focus" | "short-break" | "long-break";
 
 export default function Page() {
   // 現在のフェーズの残り時間（秒）
@@ -24,31 +28,13 @@ export default function Page() {
   // タイマーが動作中かどうか
   const [isTimerRunning, setIsTimerRunning] = useState(false);
   // 現在のフェーズ
-  const [currentPhase, setCurrentPhase] = useState<Phase>("focus");
+  const [currentPhase, setCurrentPhase] = useState<PomodoroTimerPhase>("focus");
   // 現在のサイクル
   const [currentCycle, setCurrentCycle] = useState(1);
 
-  // タイマーのカウントダウンの制御
-  useEffect(() => {
-    let interval: NodeJS.Timeout | null = null;
-
-    if (isTimerRunning && remainingTime > 0) {
-      interval = setInterval(() => {
-        setRemainingTime((prevTime: number) => prevTime - 1);
-      }, 1000);
-    } else if (remainingTime === 0) {
-      handlePhaseComplete();
-    }
-
-    return () => {
-      if (interval) clearInterval(interval);
-    };
-  }, [isTimerRunning, remainingTime]);
-
   /**
    * 現在のフェーズが完了した時の処理
-   * - 次のフェーズに移行し、適切な時間をセットする
-   * - focus → short-break → focus → ... と繰り返し、最終サイクルのshort-break後にlong-breakに移行する
+   * 次のフェーズに移行し、適切な時間をセットする
    */
   const handlePhaseComplete = () => {
     setIsTimerRunning(false);
@@ -71,11 +57,6 @@ export default function Page() {
     }
   };
 
-  /** タイマーの開始/停止を切り替える */
-  const toggleTimer = () => {
-    setIsTimerRunning(!isTimerRunning);
-  };
-
   /** タイマーをデフォルト状態にリセットする */
   const resetTimer = () => {
     setCurrentPhase("focus");
@@ -84,103 +65,22 @@ export default function Page() {
     setRemainingTime(defaultTimerConfig.focus);
   };
 
-  // 現在のフェーズをスキップする関数
-  const skipPhase = () => handlePhaseComplete();
-
-  // 表示用の分数を計算
-  const minutes = Math.floor(remainingTime / 60);
-
-  // 表示用の秒数を計算
-  const seconds = remainingTime % 60;
-
-  // プログレスバーの進捗率を計算
-  const percentage =
-    (1 -
-      remainingTime /
-        (currentPhase === "focus"
-          ? defaultTimerConfig.focus
-          : currentPhase === "short-break"
-            ? defaultTimerConfig.shortBreak
-            : defaultTimerConfig.longBreak)) *
-    100;
-
   return (
     <div className="flex flex-col items-center justify-center">
-      <div className="relative aspect-square w-80 max-w-full">
-        <svg className="h-full w-full -rotate-90" viewBox="0 0 100 100">
-          <circle
-            className="stroke-current text-gray-200"
-            strokeWidth="4"
-            cx="50"
-            cy="50"
-            r="47"
-            fill="transparent"
-          />
-          <circle
-            className={`stroke-current ${
-              currentPhase === "focus"
-                ? "text-primary"
-                : currentPhase === "short-break"
-                  ? "text-green-500"
-                  : "text-blue-500"
-            }`}
-            strokeWidth="4"
-            strokeLinecap="round"
-            cx="50"
-            cy="50"
-            r="47"
-            fill="transparent"
-            strokeDasharray="295.31"
-            strokeDashoffset={295.31 * (1 - percentage / 100)}
-          />
-        </svg>
-        <div className="absolute left-0 top-0 flex h-full w-full flex-col items-center justify-center">
-          <span className="text-5xl font-bold text-gray-800" aria-live="polite">
-            {minutes.toString().padStart(2, "0")}:
-            {seconds.toString().padStart(2, "0")}
-          </span>
-          <span
-            className="mt-2 text-lg font-medium text-gray-500"
-            aria-live="polite"
-          >
-            {currentPhase === "focus"
-              ? "集中時間"
-              : currentPhase === "short-break"
-                ? "短い休憩"
-                : "長い休憩"}
-          </span>
-          <span className="mt-1 text-sm text-gray-400">
-            {currentCycle} / {defaultTimerConfig.totalCycles}
-          </span>
-        </div>
-      </div>
-      <div className="mt-4 grid gap-4">
-        <div className="flex gap-4">
-          <Button onClick={toggleTimer}>
-            {isTimerRunning ? (
-              <>
-                <HiPause />
-                一時停止
-              </>
-            ) : (
-              <>
-                <VscDebugStart />
-                スタート
-              </>
-            )}
-          </Button>
-          <Button variant="outline" onClick={resetTimer}>
-            <RiResetLeftFill />
-            リセット
-          </Button>
-        </div>
-        <div className="grid place-content-center">
-          <Button variant="ghost" onClick={skipPhase}>
-            <BsSkipEndFill />
-            フェーズスキップ
-          </Button>
-        </div>
-      </div>
+      <PomodoroTimer
+        isTimerRunning={isTimerRunning}
+        handlePhaseComplete={handlePhaseComplete}
+        currentPhase={currentPhase}
+        currentCycle={currentCycle}
+        remainingTime={remainingTime}
+        setRemainingTime={setRemainingTime}
+      />
+      <TimerController
+        isTimerRunning={isTimerRunning}
+        setIsTimerRunning={setIsTimerRunning}
+        handlePhaseComplete={handlePhaseComplete}
+        resetTimer={resetTimer}
+      />
     </div>
   );
 }
